1- Rename "HttpApiErrorResponse" class to "ErrorResponse" 2- Move these classes (ErrorResponse, Error, GeneralError, ConcurrentModificationError) from "commercetools.Sdk.HttpApi.Domain" project to "commercetools.Sdk.Domain" project
