+++ commercetools.Sdk/commercetools.Sdk.HttpApi/commercetools.Sdk.HttpApi.csproj
@@ -34,6 +34,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\commercetools.Sdk.Client\commercetools.Sdk.Client.csproj" />
+    <ProjectReference Include="..\commercetools.Sdk.HttpApi.Domain\commercetools.Sdk.HttpApi.Domain.csproj" />
     <ProjectReference Include="..\commercetools.Sdk.Linq\commercetools.Sdk.Linq.csproj" />
     <ProjectReference Include="..\commercetools.Sdk.Serialization\commercetools.Sdk.Serialization.csproj" />
   </ItemGroup>
