+++ commercetools.Sdk/commercetools.Sdk.Serialization/JsonConverters/ErrorConverter.cs
@@ -1,8 +1,8 @@
-﻿using commercetools.Sdk.HttpApi.Domain;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
+using commercetools.Sdk.Domain.Errors;
 using Type = System.Type;
 
 namespace commercetools.Sdk.Serialization
